+++ crates/fuzz/src/transaction_executor.rs
@@ -18,17 +18,16 @@ pub struct TransactionExecutor;
 
 impl TransactionExecutor {
     #[allow(clippy::too_many_arguments)]
-    pub fn process_transaction_honggfuzz<'info, I>(
+    pub fn process_transaction_honggfuzz<I>(
         instruction_name: &str,
         client: &mut impl FuzzClient,
-        ix: &'info I,
-        snapshot: &'info mut Snapshot<'info, I>,
+        ix: &I,
         sent_txs: &mut HashMap<anchor_lang::solana_program::hash::Hash, ()>,
         config: &Config,
         accounts: &RefCell<I::IxAccounts>,
     ) -> core::result::Result<(), FuzzClientErrorWithOrigin>
     where
-        I: IxOps<'info>,
+        I: IxOps,
     {
         let program_id = ix.get_program_id();
 
@@ -37,12 +36,12 @@ impl TransactionExecutor {
             .map_err(|e| e.with_origin(Origin::Instruction(instruction_name.to_owned())))
             .expect("Accounts calculation expect");
 
+        let mut snapshot = Snapshot::new(&account_metas);
+
         let data = ix
             .get_data(client, &mut accounts.borrow_mut())
             .map_err(|e| e.with_origin(Origin::Instruction(instruction_name.to_owned())))
             .expect("Data calculation expect");
-
-        snapshot.add_metas(&account_metas);
 
         snapshot.capture_before(client).unwrap();
 
@@ -84,12 +83,7 @@ impl TransactionExecutor {
                             stats_logger.increase_successful(instruction_name.to_owned());
 
                             snapshot.capture_after(client).unwrap();
-                            let (acc_before, acc_after) = snapshot
-                                .get_snapshot()
-                                .map_err(|e| {
-                                    e.with_origin(Origin::Instruction(instruction_name.to_owned()))
-                                })
-                                .expect("Snapshot deserialization expect"); // we want to panic if we cannot unwrap to cause a crash
+                            let (acc_before, acc_after) = snapshot.get_snapshot();
 
                             if let Err(e) = ix.check(acc_before, acc_after, data).map_err(|e| {
                                 e.with_origin(Origin::Instruction(instruction_name.to_owned()))
@@ -106,8 +100,8 @@ impl TransactionExecutor {
                             stats_logger.increase_failed(instruction_name.to_owned());
                             stats_logger.output_serialized();
 
-                            let mut raw_accounts = snapshot.get_raw_pre_ix_accounts();
-                            ix.tx_error_handler(e, data, &mut raw_accounts)?
+                            let raw_accounts = snapshot.get_before();
+                            ix.tx_error_handler(e, data, raw_accounts)?
                         }
                     }
                 } else {
@@ -117,12 +111,7 @@ impl TransactionExecutor {
                     match tx_result {
                         Ok(_) => {
                             snapshot.capture_after(client).unwrap();
-                            let (acc_before, acc_after) = snapshot
-                                .get_snapshot()
-                                .map_err(|e| {
-                                    e.with_origin(Origin::Instruction(instruction_name.to_owned()))
-                                })
-                                .expect("Snapshot deserialization expect"); // we want to panic if we cannot unwrap to cause a crash
+                            let (acc_before, acc_after) = snapshot.get_snapshot();
 
                             if let Err(e) = ix.check(acc_before, acc_after, data).map_err(|e| {
                                 e.with_origin(Origin::Instruction(instruction_name.to_owned()))
@@ -132,8 +121,8 @@ impl TransactionExecutor {
                             }
                         }
                         Err(e) => {
-                            let mut raw_accounts = snapshot.get_raw_pre_ix_accounts();
-                            ix.tx_error_handler(e, data, &mut raw_accounts)?
+                            let raw_accounts = snapshot.get_before();
+                            ix.tx_error_handler(e, data, raw_accounts)?
                         }
                     }
                 }
@@ -142,17 +131,16 @@ impl TransactionExecutor {
         Ok(())
     }
     #[allow(clippy::too_many_arguments)]
-    pub fn process_transaction_afl<'info, I>(
+    pub fn process_transaction_afl<I>(
         instruction_name: &str,
         client: &mut impl FuzzClient,
-        ix: &'info I,
-        snapshot: &'info mut Snapshot<'info, I>,
+        ix: &I,
         sent_txs: &mut HashMap<anchor_lang::solana_program::hash::Hash, ()>,
         config: &Config,
         accounts: &RefCell<I::IxAccounts>,
     ) -> core::result::Result<(), FuzzClientErrorWithOrigin>
     where
-        I: IxOps<'info>,
+        I: IxOps,
     {
         let program_id = ix.get_program_id();
 
@@ -161,12 +149,12 @@ impl TransactionExecutor {
             .map_err(|e| e.with_origin(Origin::Instruction(instruction_name.to_owned())))
             .expect("Accounts calculation expect");
 
+        let mut snapshot = Snapshot::new(&account_metas);
+
         let data = ix
             .get_data(client, &mut accounts.borrow_mut())
             .map_err(|e| e.with_origin(Origin::Instruction(instruction_name.to_owned())))
             .expect("Data calculation expect");
-
-        snapshot.add_metas(&account_metas);
 
         snapshot.capture_before(client).unwrap();
 
@@ -201,12 +189,7 @@ impl TransactionExecutor {
                 match tx_result {
                     Ok(_) => {
                         snapshot.capture_after(client).unwrap();
-                        let (acc_before, acc_after) = snapshot
-                            .get_snapshot()
-                            .map_err(|e| {
-                                e.with_origin(Origin::Instruction(instruction_name.to_owned()))
-                            })
-                            .expect("Snapshot deserialization expect"); // we want to panic if we cannot unwrap to cause a crash
+                        let (acc_before, acc_after) = snapshot.get_snapshot();
 
                         if let Err(e) = ix.check(acc_before, acc_after, data).map_err(|e| {
                             e.with_origin(Origin::Instruction(instruction_name.to_owned()))
@@ -216,8 +199,8 @@ impl TransactionExecutor {
                         }
                     }
                     Err(e) => {
-                        let mut raw_accounts = snapshot.get_raw_pre_ix_accounts();
-                        ix.tx_error_handler(e, data, &mut raw_accounts)?
+                        let raw_accounts = snapshot.get_before();
+                        ix.tx_error_handler(e, data, raw_accounts)?
                     }
                 }
             }