+++ examples/fuzz-tests/arbitrary-limit-inputs-5/programs/arbitrary-limit-inputs-5/src/instructions/initialize.rs
@@ -2,7 +2,6 @@ use anchor_lang::prelude::*;
 use anchor_spl::token::{
     set_authority, transfer, Mint, SetAuthority, Token, TokenAccount, Transfer,
 };
-use trident_derive_accounts_snapshots::AccountsSnapshots;
 
 use crate::state::Escrow;
 use crate::VestingError;
@@ -62,7 +61,7 @@ pub fn _init_vesting(
     Ok(())
 }
 
-#[derive(AccountsSnapshots, Accounts)]
+#[derive(Accounts)]
 #[instruction(recipient: Pubkey)]
 pub struct InitVestingContext<'info> {
     #[account(mut)]