+++ examples/fuzz-tests/incorrect-integer-arithmetic-3/programs/incorrect-integer-arithmetic-3/src/instructions/withdraw.rs
@@ -3,8 +3,6 @@ use anchor_spl::token::{transfer, Mint, Token, TokenAccount, Transfer};
 
 use crate::{state::Escrow, VestingError};
 
-use trident_derive_accounts_snapshots::AccountsSnapshots;
-
 pub fn _withdraw_unlocked(ctx: Context<WithdrawUnlocked>) -> Result<()> {
     let escrow = &mut ctx.accounts.escrow;
 
@@ -34,7 +32,7 @@ pub fn _withdraw_unlocked(ctx: Context<WithdrawUnlocked>) -> Result<()> {
     Ok(())
 }
 
-#[derive(Accounts, AccountsSnapshots)]
+#[derive(Accounts)]
 pub struct WithdrawUnlocked<'info> {
     #[account(mut)]
     pub recipient: Signer<'info>,