+++ examples/fuzz-tests/simple-cpi-6/programs/callee/src/lib.rs
@@ -1,7 +1,5 @@
 use anchor_lang::prelude::*;
 
-use trident_derive_accounts_snapshots::AccountsSnapshots;
-
 declare_id!("HJR1TK8bgrUWzysdpS1pBGBYKF7zi1tU9cS4qj8BW8ZL");
 
 #[program]
@@ -18,7 +16,7 @@ pub mod callee {
     }
 }
 
-#[derive(Accounts, AccountsSnapshots)]
+#[derive(Accounts)]
 pub struct InitializeCallee<'info> {
     pub signer: Signer<'info>,
 }